+++ src/imio/esign/tests/test_utils.py
@@ -240,6 +240,14 @@ class TestUtils(BaseEsignTest):
         notify(ObjectModifiedEvent(annex2))
         remove_files_from_session((annex0_uid,))
         self.assertEqual(ses["size"], 6968)
+        # works also when annex deleted
+        last_annex_uid = self.uids[-1]
+        last_annex = api.content.get(UID=last_annex_uid)
+        sid, ses = add_files_to_session(signers, (last_annex_uid,))[-1]
+        self.assertEqual(ses["size"], 13982)
+        api.content.delete(last_annex)
+        self.failIf(api.portal.get_tool('portal_catalog')(UID=last_annex_uid))
+        self.assertEqual(ses["size"], 6968)
 
         # --- size-based session splitting ---
         del root_annot["imio.esign"]